MOJITO-79: Handle the case where a Bucket is already full with priority Contacts (very unlikely to happen) and you add more leafs -> old leafs will get discarded.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/dht/impl/PassiveDHTNodeRouteTable.java
@@ -99,14 +99,15 @@ class PassiveDHTNodeRouteTable extends RouteTableImpl {
     private synchronized void removeAndReplaceWithMRSCachedContact(KUID nodeId) {
         Bucket bucket = getBucket(nodeId);
         boolean removed = bucket.remove(nodeId);
-        assert (removed == true);
-        
-        Contact mrs = bucket.getMostRecentlySeenCachedContact();
-        if (mrs != null) {
-            removed = bucket.removeCachedContact(mrs.getNodeID());
-            assert (removed == true);
-            
-            bucket.addLiveContact(mrs);
+
+        if(removed) {
+            Contact mrs = bucket.getMostRecentlySeenCachedContact();
+            if (mrs != null) {
+                removed = bucket.removeCachedContact(mrs.getNodeID());
+                assert (removed == true);
+                
+                bucket.addLiveContact(mrs);
+            }
         }
     }
